# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LegaBit API",
    "description": "Versioned contracts for the LegaBit backend.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/auth/{*}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/markets": {
      "get": {
        "operationId": "getMarkets",
        "summary": "Get cryptocurrency market data",
        "tags": [
          "markets"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "usd",
                "eur",
                "btc"
              ],
              "default": "usd"
            },
            "in": "query",
            "name": "currency",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "coins": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "symbol": {
                            "type": "string",
                            "minLength": 1
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "image": {
                            "type": "string"
                          },
                          "current_price": {
                            "type": "number",
                            "nullable": true
                          },
                          "market_cap": {
                            "type": "number",
                            "nullable": true
                          },
                          "market_cap_rank": {
                            "type": "integer",
                            "minimum": 1,
                            "nullable": true
                          },
                          "price_change_percentage_24h": {
                            "type": "number",
                            "nullable": true
                          },
                          "price_change_percentage_7d_in_currency": {
                            "type": "number",
                            "nullable": true
                          },
                          "total_volume": {
                            "type": "number",
                            "nullable": true
                          },
                          "circulating_supply": {
                            "type": "number",
                            "nullable": true
                          },
                          "sparkline_in_7d": {
                            "type": "object",
                            "properties": {
                              "price": {
                                "type": "array",
                                "items": {
                                  "type": "number"
                                }
                              }
                            },
                            "required": [
                              "price"
                            ],
                            "additionalProperties": false
                          },
                          "last_updated": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "symbol",
                          "name",
                          "image",
                          "current_price",
                          "market_cap",
                          "market_cap_rank",
                          "price_change_percentage_24h",
                          "total_volume",
                          "circulating_supply",
                          "last_updated"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "global": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "active_cryptocurrencies": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "markets": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "total_market_cap": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "number"
                              }
                            },
                            "total_volume": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "number"
                              }
                            },
                            "market_cap_percentage": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "number"
                              }
                            },
                            "market_cap_change_percentage_24h_usd": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "active_cryptocurrencies",
                            "markets",
                            "total_market_cap",
                            "total_volume",
                            "market_cap_percentage",
                            "market_cap_change_percentage_24h_usd"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "data"
                      ],
                      "additionalProperties": false
                    },
                    "fetchedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "coins",
                    "global",
                    "fetchedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/newsletter/subscriptions": {
      "post": {
        "operationId": "subscribeNewsletter",
        "summary": "Subscribe to the LegaBit newsletter",
        "tags": [
          "newsletter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 7,
                    "maxLength": 32,
                    "pattern": "^[+()\\d\\s.-]+$"
                  },
                  "source": {
                    "type": "string",
                    "maxLength": 64,
                    "default": "landing"
                  }
                },
                "required": [
                  "email",
                  "phone"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "alreadySubscribed": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "alreadySubscribed",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "operationId": "getHealthLive",
        "summary": "Check process liveness",
        "tags": [
          "health"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "getHealthReady",
        "summary": "Check required dependencies",
        "tags": [
          "health"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ready"
                      ]
                    },
                    "dependencies": {
                      "type": "object",
                      "properties": {
                        "mongodb": {
                          "type": "string",
                          "enum": [
                            "up"
                          ]
                        }
                      },
                      "required": [
                        "mongodb"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "status",
                    "dependencies"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get the authenticated user",
        "tags": [
          "identity"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "name": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "image"
                      ],
                      "additionalProperties": false
                    },
                    "session": {
                      "type": "object",
                      "properties": {
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "user",
                    "session"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "VALIDATION_ERROR",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "SERVICE_UNAVAILABLE",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Service liveness and readiness"
    },
    {
      "name": "identity",
      "description": "Authenticated actor information"
    },
    {
      "name": "markets",
      "description": "Public cryptocurrency market data"
    },
    {
      "name": "newsletter",
      "description": "Public newsletter subscriptions"
    }
  ]
}